added unit tests for register, clear and logout

+++ server/src/main/java/service/LogoutService.java
@@ -2,6 +2,7 @@ package service;
 
 import dataAccess.AuthDAO;
 
+import dataAccess.BadRequestException;
 import dataAccess.UnauthorizedAccessException;
 import requests.LogoutRequest;
 
@@ -14,6 +15,9 @@ public class LogoutService {
     }
 
     public void logout() throws UnauthorizedAccessException {  //user cannot logout another user, ie AuthData can only remove itself
+        if(auth==null || authToken==null) {
+            throw new UnauthorizedAccessException("{ \"message\": \"Error: unauthorized\" }");
+        }
         auth.deleteAuth(authToken);
     }
